Revert "Fix a bug: unable to start activity when AM is killing the process."
This reverts commit 0b278659671f5de070fa38ab9c90b2d32af344c1.

I spoke to Dianne and she said that if this is making a problem go away then it is only hiding it and there is a deeper problem we need to fix.

I think if killedByApp is true then we should set app and app.thread to null. I'll look at it further when I have more time.

Change-Id: I44c29e15ca654666440472eab2d2af1c5af80b9a