Remove redundant call to window manager addTask()
If there are activities in persisted tasks then calling addAppToken()
will automatically call addTask(). If there are no activities then
calling addTask() is a nop. In either case the call to addTask() is
unnecessary. It was actually worse than unnecessary because in the
former case we ended up with two identical tasks in the Window
Manager.

Fixes bug 16958544.

Change-Id: I2dc4b50aa94668873c1a783c47e0c696d62616f0