Recompute focus stack if cleared while starting an activity.
When starting an activity with Intent.FLAG_ACTIVITY_CLEAR_TOP flag,
the activity is destoried which can also cause its task to be removed
from its current stack if the activity process record is null. We now
recompute the stack for the activity task when this occurs so we
don't NPE later on.

Bug: 19552874

Change-Id: I50f51ca6dc32d4642f78d59cae93b0774bc6cdb7
(cherry picked from commit 3f715ff97c9802d253b41f86acbfd6bc1d837069)