Don't exceed max. allowed recent document tasks for an app.
We were exceeding the maximum allowed recents document tasks entry
for an application package because the full Intent object was been
compared for equality which will be different for each task. Now
we are only comparing that the task components match which is
sufficient to determining that the task belongs to the same
application package.

Bug: 18642190
Change-Id: I831b0fcffb876d51e9439e7a7a4c75bb0881db7c