Do not ensure visibility when just resizing
No need to ensure the visibilities of activities if only the
bounds change. If we resize the docked stack to fullscreen this
method still gets called.

This makes multi-window dogfoodable again with all that DEBUG
churn we have enabled, but also helps performance in general
when the history is quite big.

Bug: 25015474
Change-Id: I9ae0281392353e6c01563d021bca7dc23125b14f