Clean-up starting window if resuming activity in another stack
We currently remove starting windows for initializing activities if we are
resuming another activity in the same stack. However, the starting window
can be orphaned if we are resuming an activity in a different stack and
start another activity in the stack with the starting window in the paused
state at the same time.
We now remove starting windowing for initializing activities that aren't
the top activity in the stack regardless of the stack we are resuming an
activity in.

Bug: 26659857
Change-Id: I6081d52d0c4c0d3556201346cfda7a38e4d02c8e