Consolidate the HealthStats UI to only use milliseconds.
BatteryStats uses a mix of microseconds and milliseconds, and keeping
which one is which straight is difficult.  The internal bookkeeping
is left using us where it already does, for the extra precision. But
having the API be mixed will only make it harder for developers, and
lead them to make the same mistake that I did in the API where I
missed some conversions.

Bug: 28197858
Change-Id: I99114bae259b9bdd47ce5c22e724d87bbd63336d