Finish immediately if the resumed activity is already visible
When in FINISH_AFTER_VISIBLE mode, we need to check visibility of
both the current activity, and the activity to be resumed. If the
resumed activity is already visible, we should finish immediately.
We can't put it into stopping state and wait because there won't be
a window visible callback from the next activity.

bug: 28378359

Change-Id: Ib1fb44acfec33616da76f48102cc00b453e41914